Removed the redundant PlayerMove script
