+++ TPC/Assets/Scenes/Main.unity
@@ -408,8 +408,8 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 1018689113}
-  - component: {fileID: 1018689115}
   - component: {fileID: 1018689114}
+  - component: {fileID: 1018689115}
   m_Layer: 0
   m_Name: Player
   m_TagString: Untagged
@@ -460,10 +460,9 @@ MonoBehaviour:
   m_GameObject: {fileID: 1018689112}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 047079bd310dcd64fba557be66e9045f, type: 3}
+  m_Script: {fileID: 11500000, guid: 1d56ae9357a650e46bf3ae8e64ed1b95, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  mCharacterController: {fileID: 1018689114}
   mAnimator: {fileID: 1365547314}
   mWalkSpeed: 1.5
   mRotationSpeed: 25
